backend/pyproject.toml
[tool.poetry]
name = "incident"
version = "0.1.0"
description = ""
authors = ["Andrey Meshkov <a.miashkov@jsglobal.by>"]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.2.8,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
django-filter = ">=25.2,<26.0"
drf-spectacular = ">=0.29.0,<0.30.0"
django-cors-headers = ">=4.9.0,<5.0.0"
psycopg2-binary = "^2.9.11"
django-split-settings = "^1.3.2"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^6.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

